API Reference

Configuration

Umbrella.Configuration.Options — Type

Options(client_id, client_secret, redirect_uri, success_redirect, failure_redirect, scopes, state, providerOptions)

Represents an Configuration Options with fields:

client_id::String Client id from an OAuth 2 provider

client_secret::String Secret from an OAuth 2 provider

redirect_uri::String Determines where the API server redirects the user after the user completes the authorization flow

success_redirect::String URL path when OAuth 2 successed

failure_redirect::String URL path when OAuth 2 failed

scope::String OAuth 2 scopes

state::String Specifies any string value that your application uses to maintain state between your authorization request and the authorization server's response

Umbrella.init — Function

init(type::Symbol, config::Configuration.Options)

Initiate an OAuth 2 instance for the given provider.

Umbrella.register — Function

register(type::Symbol, oauth2_actions::OAuth2Actions)

Register a newly implemented OAuth 2 provider.

Handlers

Umbrella.OAuth2 — Type

OAuth2(type, redirect, token_exchange)

Represents an OAuth2 with fields:

type::Symbol OAuth 2 provider

redirect::Function Generates the redirect URL and redirects users to provider's OAuth 2 server to initiate the authentication and authorization process.

token_exchange::Function Use code responded by the OAuth 2 server to exchange an access token, and get user profile using the access token.

redirect and token_exchange functions are produced by init function, they are configured for a specific OAuth 2 provider, example usage:

const options = Configuration.Options(
    # fill in the values
)
const oauth2 = Umbrella.init(:google, options)

# perform redirect
oauth2.redirect()

# handle token exchange
oauth2.token_exchange(code, verify)

OAuth 2 Providers